refactor: [TEO-281] move download access request interface to access request module
