+++ backend/src/usecases/access-request/download-access-requests.usecase.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { IAccessRequestDownload } from 'src/common/interfaces/access-request-download.interface';
+import { IAccessRequestDownload } from 'src/modules/access-request/interfaces/access-request-download.interface';
 import { IUseCaseService } from 'src/common/interfaces/use-case-service.interface';
 import { AccessRequestStatus } from 'src/modules/access-request/enums/access-request-status.enum';
 import { FindManyAccessRequestsOptions } from 'src/modules/access-request/model/access-request.model';
